+++ sdk/eventgrid/azure-eventgrid/azure/eventgrid/_helpers.py
@@ -191,4 +191,4 @@ def _get_json_content(obj):
                 + "body follows the EventGridEvent schema."
                 )
         except: # pylint: disable=bare-except
-            return obj
+            return json.loads(obj)